backend/requirements.txt
asgiref
Django
django-cors-headers
django-extensions
djangorestframework
Markdown
pydot
pyparsing
pytz
six
sqlparse